refactor: csm gate tree extra validation text
